+++ src/main/java/dev/easbarba/onur/actions/Pull.java
@@ -13,27 +13,12 @@
 *  along with Onur. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package dev.easbarba.onur.misc;
+package dev.easbarba.onur.actions;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.nio.file.Paths;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-public final class GlobalsTest {
-    private Globals globals;
-
-    @BeforeEach
-    void init() {
-        this.globals = Globals.getInstance();
+public class Pull implements IAction {
+    @Override
+    public void run() {
+        throw new UnsupportedOperationException("Unimplemented method 'run'");
     }
 
-    @Test
-    public void shouldHaveValidFilesExactly() {
-        assertEquals(
-                Paths.get(System.getProperty("user.home").toString(), ".config", "onur").toAbsolutePath().toString(),
-                globals.get("config-home").toString());
-    }
 }