+++ src/test/java/dev/easbarba/onur/database/ConfigurationsTest.java
@@ -13,27 +13,35 @@
 *  along with Onur. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package dev.easbarba.onur.misc;
+package dev.easbarba.onur.database;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.nio.file.Paths;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public final class GlobalsTest {
-    private Globals globals;
+public final class ConfigurationsTest {
+    private Configurations configurations;
 
     @BeforeEach
     void init() {
-        this.globals = Globals.getInstance();
+        this.configurations = new Configurations();
     }
 
     @Test
-    public void shouldHaveValidFilesExactly() {
-        assertEquals(
-                Paths.get(System.getProperty("user.home").toString(), ".config", "onur").toAbsolutePath().toString(),
-                globals.get("config-home").toString());
+    public void shouldHaveCorrectConfigurationsSize() {
+        assertEquals(this.configurations.count(), 2);
+    }
+
+    @Test
+    public void shouldHaveASpecificConfiguration() {
+        assertEquals(this.configurations.namespath().stream().findFirst().get().getFileName().toString(),
+                "etc.json");
+    }
+
+    @Test
+    public void shouldConfigurationExist() {
+        assertTrue(this.configurations.exists());
     }
 }